feature: compute total 'mass' by category
we consider the total area covered by dots as total mass of Fauna (without considering that the dots can overlap).  we then compute what percentage of the mass belongs to what category; in this case, we only consider color.  in the future this will be different types of Fauna.  additionally, this revision makes some improvements on our testing classes so that a similar initalization can be used for multiple tests